Fix compilation and build issues for OpenPANA
- Add missing preprocessor macro definitions for ISSERVER/ISCLIENT/ISPRE
- Fix type mismatches in EAP function calls (size_t vs unsigned int)
- Add missing includes for semaphore.h and config.h
- Resolve multiple definition errors by converting globals to extern declarations
- Fix library linking order and circular dependencies with --start-group/--end-group
- Remove unused acinclude.m4 file
- Add comprehensive Makefile with separate object files for server/client/relay builds

All three binaries (openpaa, openpac, openpre) now build successfully.

+++ src/state_machines/paamachine.c
@@ -751,7 +751,7 @@ int newKeyAvailable() {
 	//Tries to retrieve a Master Session Key (MSK) from the EAP entity
 		if (eapKeyAvailable == TRUE) {
 			pana_debug("EAP lower-layer Key Available");
-			unsigned int key_len;
+			size_t key_len;
 			#ifdef ISCLIENT
 			u8* key = eap_peer_get_eapKeyData(&(current_session->eap_ctx), &key_len);
 			#endif
